refact : GlobalStyle -> component 분리
- 글로벌 상태에서 기본적인 공용 컴포넌트를 일부 관리
- shard/component/[ 기능 ]  해당 되는 컴포넌트로 분류하여 Base로 설정
