+++ client/src/shared/components/button/index.ts
@@ -1,0 +1,1 @@
+export { BaseButton } from './Button';
