+++ client/src/shared/components/card/index.ts
@@ -1,0 +1,1 @@
+export { BaseCard } from './Card';
