+++ client/src/shared/components/textarea/index.tsx
@@ -1,0 +1,1 @@
+export { BaseArea } from './TextArea';
