+++ client/src/shared/components/title/index.ts
@@ -1,0 +1,1 @@
+export { BaseTitle } from './MainTitle';
