+++ client/src/widgets/header/data/index.tsx
@@ -27,6 +27,11 @@ export const navList: MenuItemType[] = [
     path: '/qna/list',
     icon: <BiCommentDetail />,
   },
+  {
+    title: 'admin',
+    path: '/admin',
+    icon: <BiEdit />,
+  },
 ];
 
 export const dropMenuList: DropItemType[] = [
